Save banned time within the DB and make sure bans are saved
If PyFilter was to close after some IP's have been banned, which were not saved via iptables, when iptables is reloaded that IP will be saved in the database, and not banned within the DB causing quite an issue, this way, if PyFilter was to close it will save the outstanding bans to the file ready for reload. The time the IP was banned is now saved within redis or sqlite.